separate V1 and V2 interpretors in two different classes
